+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.1.4
+
+### New features 
+
+* Database name can now contain multiple schemas in UI field, which tells the driver to scan selected databases. (separator can be set in `metabase.driver.clickhouse/SEPARATOR`) 
+
 # 1.1.2
 
 ### Bug fixes